Break implementation generators out into separate classes.
